+++ src/games/voxel-realms/store/traits.ts
@@ -1,4 +1,0 @@
-import { trait } from "koota";
-import { createInitialVoxelState } from "../engine/voxelSimulation";
-
-export const VoxelTrait = trait(() => createInitialVoxelState());